Reduce the maximum preprocessor configs to 1.
The cppcheck --max-configs parameter controls how many different
configurations of defines to analyze.  By default, it will
try up to 12.  The problem is that this causes the analysis
to take quite a long time to run.  Instead, force the maximum
configs to 1, which trades off a bit of coverage for a
massive increase in speed (more than twice as fast in
my testing).  Since we run cppcheck on multiple platforms
(i.e. Windows, Linux, and macOS), we'll get a lot of coverage
anyway, just not all on one platform.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>