Add SPDX identifiers to the licenses.
This is so we can use them at a higher layer.  Note that the
SPDX identifier for "BSD License 2.0" is made up, since that
license is also made up.  We eventually need to remove it,
but some of the core packages still depend on it (and really
should be switched over to BSD-3-Clause).

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>