[ament_cpplint] Process errors without linenums
It is possible for cpplint to output errors that do not have an
associated line number. The xunit file generation process throws a
TypeError when dealing with such errors since it strictly expects an
integral value for the line number (and not a NoneType).

This commit lets ament_cpplint process errors without line numbers,
replacing the integral line number value with 'None' in the output xunit
file when it is appropriate to do so.

Signed-off-by: Abrar Rahman Protyasha <aprotyas@u.rochester.edu>