Declare dependencies on expected flake8 plugins.
These flake8 plugins have been in the from-source setup instructions for
ROS 2 a long time. For Ubuntu 22.04 we finally got all of them packaged
so we can start removing pip packages from our installation
requirements.

I think this change should be back-portable to Humble but it can't go
back to Focal because not all of these packages are available there.

Signed-off-by: Steven! Ragnarök <steven@nuclearsandwich.com>