Improve ament_copyright performance drastically.
While running some repeated tests locally, I was finding
that ament_copyright was taking about 3 seconds to process
a small number of files (34).  After looking into it, I
found that it was repeatedly calling get_licenses() and
get_copyright_names().

However, we notice that get_licenses() and get_copyright_names()
are pulling the known copyrights from the package itself.
Thus, there is no way that these can change during the runtime
of the program.

This PR changes it so that we only call these functions once
in main(), and the rest of the code just uses that.  With
this in place, the runtime for the 34 files I was checking
went from 3 seconds down to 0.2 seconds.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>